test: move to examples; add point_discharge2d
