+++ test/examples/burgers.py
@@ -19,6 +19,7 @@ fields = ['uv_2d']
 end_time = 0.5
 dt = 1/n
 dt_per_export = 2
+steady = False
 
 
 def get_function_spaces(mesh):
