+++ test/examples/solid_body_rotation_split.py
